stream-analytics/requirements.txt
# Kafka and Avro dependencies
confluent-kafka==2.1.1
fastavro==1.5.4
avro-python3==1.10.2

# Spark dependencies
pyspark==3.3.1
delta-spark==2.3.0

# Data processing dependencies
pandas==1.5.3
numpy==1.24.3
scipy==1.10.1

# Airflow dependencies
apache-airflow==2.5.1
apache-airflow-providers-apache-spark==4.0.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-http==4.1.0

# Data generation and utilities
faker==18.13.0
python-dateutil==2.8.2
tenacity==8.2.2
schedule==1.2.0
pydantic==1.10.8
uuid==1.30

# Configuration and serialization
pyyaml==6.0
jsonschema==4.17.3

# HTTP and API
requests==2.28.2
flask==2.3.2
connexion==2.14.2

# Monitoring and logging
prometheus-client==0.16.0
pythonjsonlogger==2.0.7

# Testing
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0

# Code quality
black==23.3.0
flake8==6.0.0
isort==5.12.0

# AWS S3 integration
boto3==1.26.133
s3fs==2023.4.0

# Development tools
ipython==8.13.2
jupyter==1.0.0
notebook==6.5.4

# Database
alembic==1.13.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9 